pyproject.toml
[project]
name = "django-q"
version = "0.1.0"
description = "Django application with django-q2 task queue, REST API, and Swagger documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.4",
    "django-q2>=1.9.0",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.27.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "mjml>=0.12.0",
    "html2text>=2024.2.26",
]
